refactor(gateway): extract transient-error helpers from run.py (slice 5 of #54962)
Fifth slice of the gateway god-file unpacking: extract
_is_transient_network_error and _gateway_loop_exception_handler into
gateway/error_helpers.py.

- Byte-identical extraction (AST-verified against origin/main): the
  classifier walks the exception cause chain (#31066/#31110); the loop
  handler wires it into the event-loop safety net
- gateway/run.py: -76 lines; helpers imported at the extraction point
- 8 tests pass (loop exception handler)

Follow-up to #77433 (slice 1), #77438 (slice 2), #77452 (slice 3),

#77450 (slice 4). Progress on #54962.
